move kill_java_player_processes to mod
