+++ src/controllers/matchmakers/classic2v2.rs
@@ -5,7 +5,7 @@ use std::{
     fs::{self, File},
     io::{self, BufRead, BufReader},
     path::Path,
-    process::{Command, ExitStatus, Output, Stdio},
+    process::{Command, ExitStatus, Stdio},
     sync::{Arc, Mutex},
     thread,
     time::Duration,
@@ -130,62 +130,6 @@ pub fn run_round(competition_id: String) -> Result<(), MatchMakerError> {
         return Err(MatchMakerError::DatabaseError(e));
     }
     println!("Competition done!");
-    Ok(())
-}
-
-/// Kill all processes running with the command "java Player."
-pub fn kill_java_player_processes() -> Result<(), std::io::Error> {
-    // Get a list of all processes with "java Player" in their command line
-    let ps_output = Command::new("ps").arg("ax").output()?;
-
-    // Convert the output to a string
-    let ps_output_str = String::from_utf8_lossy(&ps_output.stdout);
-
-    // Split the output into lines
-    let process_lines: Vec<&str> = ps_output_str.lines().collect();
-
-    // Iterate through the lines and find processes with "java Player"
-    for process_line in process_lines {
-        if process_line.contains("java Player") {
-            // Extract the process ID (PID)
-            let pid_str = process_line.split_whitespace().next().unwrap_or_default();
-
-            // Parse the PID as an integer
-            if let Ok(pid) = pid_str.parse::<i32>() {
-                // Kill the process using the "kill" command
-                let kill_result = Command::new("kill")
-                    .arg("-9") // Use SIGKILL to forcefully terminate the process
-                    .arg(pid.to_string())
-                    .output();
-
-                match kill_result {
-                    Ok(Output {
-                        status,
-                        stdout,
-                        stderr,
-                    }) => {
-                        if status.success() {
-                            println!(
-                                "Killed process with PID {}: {:?}",
-                                pid,
-                                String::from_utf8_lossy(&stdout)
-                            );
-                        } else {
-                            eprintln!(
-                                "Failed to kill process with PID {}: {:?}",
-                                pid,
-                                String::from_utf8_lossy(&stderr)
-                            );
-                        }
-                    }
-                    Err(e) => {
-                        eprintln!("Error killing process with PID {}: {:?}", pid, e);
-                    }
-                }
-            }
-        }
-    }
-
     Ok(())
 }
